Move actions and HandlersRegistry under event_handlers/ directory
